# project/doc/swagger.yaml
openapi: 3.0.1
info:
  title: PDF-PHP API
  description: Eine Schnittstellendokumentation für das PDF-PHP Projekt
  version: 0.0.1
servers:
  - url: https://pdf-php.ddev.site/
    description: DDEV Umgebung
  - url: https://pdf.katschek.de
    description: Der Zukünftige Staging Server
paths:
  /v1/json/document/template:
    post:
      summary: PDF generierung
      description: Generiert ein PDF auf Basis des gegebenen templates, welches ein json format haben muss.
      requestBody:
        description: Das json template.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PdfTemplate'
      responses:
        '200':
          description: Ein json string, mit einem array => value mapping, welches gegen die Schnittstelle für die Erstellung gegeben werden muss.
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    PdfTemplate:
      required:
        - author
        - filename
      type: object
      properties:
        author:
          type: string
          description: Der Author von der PDF Datei.
          example: Ruwen Katschek
        pages:
          type: array
          items:
            $ref: '#/components/schemas/Page'
        filename:
          type: string
          description: PDF File name
          example: rechnung.pdf
    Page:
      type: object
      properties:
        elements:
          type: array
          items:
            $ref: '#/components/schemas/PageElement'
    PageElement:
      type: object
      properties:
        cellWidth:
          type: integer
          description: Die Höhe einer Zelle
          example: 12
        cellHeight:
          type: integer
          description: Die Breite einer Zelle
          example: 23
        xCellPosition:
          type: integer
          description: Die X Position der Zelle
          example: 12
        yCellPosition:
          type: integer
          description: Die Y Position der Zelle
          example: 12
        elementName:
          type: string
          description: Der Name des Elements
          example: Datum
          required: true
        value:
          type: string
          description: Das ist der Wert der Zelle. Wenn der Wert leer gelassen wird (null) wird ein platzhalter erstellt um dynamisch inhalt zu laden. Wenn ein Wert gesetzt wird, dann ist dieser statisch in dem Template
          example: 12.04.1996
          required: false